requirements.txt
alembic==1.7.3
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
python-dotenv==0.19.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.6.0

PyJWT~=2.1.0
slugify~=0.0.1